+++ firmware/hdl/packages/gt_mp7_core_pkg.vhd
@@ -8,11 +8,12 @@
 -- v1.15.2: Added tcl script for "manualy" bit file generation (after timing errors)
 --
 -- frame: v1.2.4 (see frame.vhd)
--- gtl: v1.16.1 (see gtl_module_tpl.vhd)
+-- gtl: v1.17.0 (see gtl_module_tpl.vhd)
 -- fdl: v1.3.6 (see fdl_module.vhd)
 
 -- gtl history:
--- HB 2021-10-19: v1.17.0: Implemented logic for jet DISP (displaced) cut [DISP = bit 27 of jet data] in calo_comparators.vhd.
+-- HB 2021-11-12: v1.17.0: Implemented logic for jet DISP (displaced) cut [DISP = bit 27 of jet data] in calo_comparators.vhd.
+-- HB 2021-10-27: v1.16.2: Bug fix in matrix_corr_cond.vhd.
 -- HB 2021-10-18: v1.16.1: Bug fix in calo_comparators.vhd.
 -- HB 2021-08-31: v1.16.0: Changed logic for ROMs (mass over DR) [regenerated IPs for ROMs].
 -- HB 2021-06-10: v1.15.1: Added hadronic shower triggers.