+++ CHANGELOG.md
@@ -4,7 +4,7 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
-## [v1.17.0] - 2021-10-19
+## [v1.17.0] - 2021-11-12
 ### Comment
 
 - mp7_ugt firmware release v1.17.0
@@ -23,6 +23,18 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 ### Added
 - source files:
   - ../payload/gtl/common/pt_disp_comp.vhd
+
+## [v1.16.2] - 2021-10-27
+### Comment
+
+- mp7_ugt firmware release v1.16.2
+- bug fixed in matrix_corr_cond.vhd
+
+### Changed
+- source files:
+  - ../payload/gtl/common/matrix_corr_cond.vhd
+  - ../payload/gtl_module_tpl.vhd
+  - ../packages/gt_mp7_core_pkg.vhd
 
 ## [v1.16.1] - 2021-10-18
 ### Comment